fix bestshare type and parse datetime instead of unix timestamp

+++ src/Models/Pool/PoolStatistics.cs
@@ -5,7 +5,7 @@ public class PoolStatistics
     public double Diff { get; set; }
     public int Accepted { get; set; }
     public int Rejected { get; set; }
-    public int BestShare { get; set; }
+    public double BestShare { get; set; }
     public double SPS1m { get; set; }
     public double SPS5m { get; set; }
     public double SPS15m { get; set; }
